fixin problem by conditions of the task
