--- product-by-id-service/products.js
@@ -1,68 +1,0 @@
-const productsList = [
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4891",
-        name: "Nike Men's Stroke Running Shoe CX",
-        brand: "Nike",
-        size: "41",
-        color: "black",
-        price: "49"
-    },
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4892",
-        name: "Nike Men's Stroke Running Shoe DS",
-        brand: "Nike",
-        size: "42",
-        color: "blue",
-        price: "50"
-    },
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4893",
-        name: "Nike Men's Stroke Running Shoe FR",
-        brand: "Nike",
-        size: "43",
-        color: "blue",
-        price: "60"
-    },
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4894",
-        name: "Nike Men's Stroke Running Shoe EW",
-        brand: "Nike",
-        size: "44",
-        color: "blue",
-        price: "70"
-    },
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4895",
-        name: "Nike Men's Stroke Running Shoe FQ",
-        brand: "Nike",
-        size: "45",
-        color: "blue",
-        price: "80"
-    },
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4896",
-        name: "Nike Men's Stroke Running Shoe TR",
-        brand: "Nike",
-        size: "46",
-        color: "blue",
-        price: "90"
-    },
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4897",
-        name: "Nike Men's Stroke Running Shoe GH",
-        brand: "Nike",
-        size: "47",
-        color: "blue",
-        price: "100"
-    },
-    {
-        id: "99039f46-7543-42be-b075-9ba8cfcd4898",
-        name: "Nike Men's Stroke Running Shoe GR",
-        brand: "Nike",
-        size: "48",
-        color: "blue",
-        price: "110"
-    }
-];
-
-module.exports = { productsList };
