+++ product-service/serverless.yml
@@ -61,6 +61,12 @@ functions:
       - http:
           path: products
           method: get
+  getProductById:
+    handler: handler.getProductById
+    events:
+      - http:
+          path: /products/{productId}
+          method: get  
 #    The following are a few example events you can configure
 #    NOTE: Please make sure to change your handler code to work with those events
 #    Check the event documentation for details
